move get_region_base_address_offset to break circular import
