+++ spinn_front_end_common/utility_models/data_speed_up_packet_gatherer_machine_vertex.py
@@ -27,8 +27,6 @@ from spinnman.messages.sdp import SDPMessage, SDPHeader, SDPFlag
 from spinnman.messages.scp.impl.iptag_set import IPTagSet
 from spinnman.connections.udp_packet_connections import SCAMPConnection
 from spinnman.model.enums.cpu_state import CPUState
-from spinn_front_end_common.utilities.utility_calls import (
-    get_region_base_address_offset)
 from pacman.executor.injection_decorator import inject_items
 from pacman.model.graphs.common import EdgeTrafficType
 from pacman.model.graphs.machine import MachineVertex
@@ -39,7 +37,8 @@ from spinn_front_end_common.interface.provenance import ProvenanceWriter
 from spinn_front_end_common.utilities.globals_variables import (
     get_simulator)
 from spinn_front_end_common.utilities.helpful_functions import (
-    convert_vertices_to_core_subset, n_word_struct)
+    convert_vertices_to_core_subset, get_region_base_address_offset,
+    n_word_struct)
 from spinn_front_end_common.utilities.emergency_recovery import (
     emergency_recover_state_from_failure)
 from spinn_front_end_common.abstract_models import (
